+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix issue [#142](https://github.com/intersystems/language-server/issues/142): Don't suggest UDL keyword in code completion if it's already specified
 - Fix issue [#143](https://github.com/intersystems/language-server/issues/143): Don't create Folding Ranges for labels in classes
 - Fix issue [#144](https://github.com/intersystems/language-server/issues/144): No intellisense for back-to-back macros
+- Fix issue [#145](https://github.com/intersystems/language-server/issues/145): Compiler keyword completion list includes invalid keywords
 
 ## [1.1.2] - 2021-02-23
 - Fix issue [#137](https://github.com/intersystems/language-server/issues/137): Completion request fails for local variables